pyproject.toml
[project]
name = "voice-judge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.16.1",
    "openai>=1.82.0",
    "weave>=0.51.48",
]
